# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Dify2OpenAI","version":"0.1.0"},"paths":{"/are-you-ok":{"get":{"summary":"Are You Ok","operationId":"are_you_ok_are_you_ok_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AreYouOk"}}}}}}},"/v1/models":{"get":{"tags":["OpenAI"],"summary":"Get Models","operationId":"get_models_v1_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelsResponse"}}}}},"security":[{"APIKeyHeader":[]}]}},"/v1/chat/completions":{"post":{"tags":["OpenAI"],"summary":"Chat Completions","operationId":"chat_completions_v1_chat_completions_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatCompletionsBody"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"AreYouOk":{"properties":{"message":{"type":"string","title":"Message","description":"How are you?","default":"I'm OK!"}},"type":"object","title":"AreYouOk"},"ChatCompletionsBody":{"properties":{"type":{"anyOf":[{"type":"string","enum":["chat","agent","workflow"]},{"type":"null"}],"title":"Type","description":"应用类型"},"model":{"type":"string","title":"Model","description":"要使用的模型的 ID","example":"dify"},"conversation_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Conversation Id","description":"会话 ID，需要基于之前的聊天记录继续对话，必须传之前消息的 conversation_id。"},"auto_generate_name":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Auto Generate Name","description":"自动生成标题，默认 true。 若设置为 false，则可通过调用会话重命名接口并设置 auto_generate 为 true 实现异步生成标题。","default":true},"inputs":{"type":"object","title":"Inputs","description":"允许传入 App 定义的各变量值。 inputs 参数包含了多组键值对（Key/Value pairs），每组的键对应一个特定变量，每组的值则是该变量的具体值","default":{}},"user":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User","description":"用户标识，用于定义终端用户的身份，方便检索、统计。 由开发者定义规则，需保证用户标识在应用内唯一。如果不传默认随机生成一个 UUID","default":"e8f5a6ee-5b1d-44be-b72a-efb731e87b57"},"detail":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Detail","description":"是否返回详细信息（usage）","default":false},"stream":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Stream","description":"是否使用流式模式返回","default":false},"messages":{"items":{"$ref":"#/components/schemas/Message"},"type":"array","maxItems":1,"minItems":1,"title":"Messages","description":"消息列表"}},"type":"object","required":["model","messages"],"title":"ChatCompletionsBody"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Message":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"Message"},"Model":{"properties":{"id":{"type":"string","title":"Id"},"object":{"type":"string","title":"Object"},"created":{"type":"integer","title":"Created"},"owned_by":{"type":"string","title":"Owned By"}},"type":"object","required":["id","object","created","owned_by"],"title":"Model"},"ModelsResponse":{"properties":{"object":{"type":"string","title":"Object"},"data":{"items":{"$ref":"#/components/schemas/Model"},"type":"array","title":"Data"}},"type":"object","required":["object","data"],"title":"ModelsResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"Authorization"}}}}